Created snake game using turtle graphics library
Created snake game which was first introduced in the Nokia 3310 mobile. I used the turtle graphics library to design the GUI of the game.  Implemented the concepts of Object Oriented Programming and Class inheritance to develop the game mechanics such as snake and its elongation, random food generation and game over rules such as collision with walls and snake itself.